+++ lib/ksef/requests/testdata/limits_context_session_handler.rb
@@ -11,8 +11,15 @@ module KSEF
 
         # Set test context session limits
         # @param limits_data [Hash] Limits configuration
-        # @option limits_data [Integer] :max_sessions Maximum number of sessions
-        # @option limits_data [Integer] :max_invoices_per_session Maximum invoices per session
+        # @option limits_data [Hash] :online_session Online session limits
+        # @option limits_data [Hash] :batch_session Batch session limits
+        #
+        # Session limits hash can contain:
+        # - maxInvoiceSizeInMB [Integer] Max invoice size in MB (RC5.3+, replaces maxInvoiceSizeInMib)
+        # - maxInvoiceWithAttachmentSizeInMB [Integer] Max invoice with attachment size in MB (RC5.3+)
+        # - maxInvoiceSizeInMib [Integer] (deprecated, will be removed 2025-10-27)
+        # - maxInvoiceWithAttachmentSizeInMib [Integer] (deprecated, will be removed 2025-10-27)
+        #
         # @return [Hash] Limits response
         def call(limits_data:)
           response = @http_client.post("testdata/limits/context/session", body: limits_data)